requirements.txt
# Core ML
torch>=2.0.0
open_clip_torch
Pillow
faiss-cpu

# Data
pandas
numpy
scikit-learn

# LangChain
langchain
langchain-community
langchain-openai
langchain-core

# NLP Metrics
nltk
rouge-score

# API & Environment
python-dotenv

# Demo UI
gradio
